requirements.txt
# Web Framework
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# NLP Vietnamese
underthesea==6.7.0
pyvi==0.1.1
vncorenlp==1.0.3

# Machine Learning
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1

# Web Scraping
scrapy==2.11.0
selenium==4.15.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0

# Database
pymongo==4.6.0
python-dotenv==1.0.0

# Utils
python-dateutil==2.8.2
tqdm==4.66.1